@@ -1,7 +1,7 @@
 {
-  "name": "electron-react-parcel-boilerplate",
-  "version": "0.7.7",
-  "description": "A boilerplate for Electron + React + Parcel.",
+  "name": "csm2",
+  "version": "2.0.0",
+  "description": "Cash shop manager",
   "main": "app/main/.parcel/main.js",
   "scripts": {
     "start": "run-s main:build watch",
@@ -19,11 +19,11 @@
     "test:jest": "jest",
     "lint": "tslint app/**/*.ts{,x}"
   },
-  "author": "shamofu",
+  "author": "now",
   "license": "MIT",
   "build": {
-    "productName": "ElectronReactParcel",
-    "appId": "org.shamofu.ElectronReactParcel",
+    "productName": "csm",
+    "appId": "gg.now.csm2",
     "files": [
       "app/**",
       "!app/main/main.ts",
@@ -77,32 +77,34 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/preset-env": "^7.4.5",
+    "@babel/core": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
     "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.3.3",
-    "@types/jest": "^24.0.15",
-    "@types/react": "^16.8.22",
-    "@types/react-dom": "^16.8.4",
+    "@babel/preset-typescript": "^7.6.0",
+    "@types/jest": "^24.0.18",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^24.8.0",
-    "electron": "^5.0.5",
-    "electron-builder": "^20.44.4",
-    "jest": "^24.8.0",
-    "nodemon": "^1.19.1",
+    "babel-jest": "^24.9.0",
+    "electron": "^6.0.11",
+    "electron-builder": "^21.2.0",
+    "jest": "^24.9.0",
+    "nodemon": "^1.19.3",
     "npm-run-all": "^4.1.5",
     "parcel-bundler": "^1.12.3",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.1",
-    "react-test-renderer": "^16.8.6",
-    "spectron": "^7.0.0",
-    "tslint": "^5.18.0",
-    "tslint-react": "^4.0.0",
-    "typescript": "^3.5.2"
+    "spectron": "^8.0.0",
+    "tslint": "^5.20.0",
+    "typescript": "^3.6.3"
   },
   "dependencies": {
     "app-root-path": "^2.2.1",
-    "electron-is-dev": "^1.1.0"
-  }
+    "crayon": "^4.1.6",
+    "crayon-svelte": "^1.0.10",
+    "edge": "^7.10.1",
+    "electron-edge-js": "^12.4.0",
+    "electron-is-dev": "^1.1.0",
+    "parcel-plugin-svelte": "^4.0.4",
+    "svelte": "^3.12.1"
+  },
+  "browserslist": [
+    "last 1 chrome versions"
+  ]
 }
